+++ src/frequenz/client/microgrid/_client.py
@@ -17,16 +17,15 @@ from frequenz.api.common import components_pb2, metrics_pb2
 from frequenz.api.microgrid import microgrid_pb2, microgrid_pb2_grpc, sensor_pb2
 from frequenz.channels import Receiver
 from frequenz.client.base import channel, client, retry, streaming
+from frequenz.client.common.enum_proto import enum_from_proto
 from frequenz.client.common.microgrid import MicrogridId
-from frequenz.client.common.microgrid.components import ComponentId
+from frequenz.client.common.microgrid.components import ComponentCategory, ComponentId
 from frequenz.client.common.microgrid.sensors import SensorId
 from google.protobuf.empty_pb2 import Empty
 from typing_extensions import override
 
 from ._component import (
     Component,
-    ComponentCategory,
-    component_category_from_protobuf,
     component_metadata_from_protobuf,
     component_type_from_protobuf,
 )
@@ -187,7 +186,7 @@ class MicrogridApiClient(client.BaseApiClient[microgrid_pb2_grpc.MicrogridStub])
         result: Iterable[Component] = map(
             lambda c: Component(
                 ComponentId(c.id),
-                component_category_from_protobuf(c.category),
+                enum_from_proto(c.category, ComponentCategory),
                 component_type_from_protobuf(c.category, c.inverter),
                 component_metadata_from_protobuf(c.category, c.grid),
             ),
@@ -386,9 +385,13 @@ class MicrogridApiClient(client.BaseApiClient[microgrid_pb2_grpc.MicrogridStub])
             raise ValueError(f"Unable to find {component_id}") from exc
 
         if comp.category != expected_category:
+            cat = (
+                comp.category.name
+                if isinstance(comp.category, ComponentCategory)
+                else comp.category
+            )
             raise ValueError(
-                f"{component_id} is a {comp.category.name.lower()}"
-                f", not a {expected_category.name.lower()}."
+                f"{component_id} has category {cat}, but {expected_category.name} was expected"
             )
 
     async def meter_data(  # noqa: DOC502 (ValueError is raised indirectly by _expect_category)